pyproject.toml
[tool.poetry]
name = "template-backend"
version = "0.1.0"
description = ""
authors = ["Mateus Santos <santos.mateus@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask-SQLAlchemy = "^2.5.1"
connexion = {extras = ["swagger-ui"], version = "^2.14.0"}
python-dotenv = "^0.20.0"
Flask-Cors = "^3.0.10"
Flask-Migrate = "^3.1.0"
pycardano = "0.7.0"
psycopg2-binary = "^2.9.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"
retry = "^0.9.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pytest-cov = "^4.0.0"
pytest-xdist = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
